Move Callable.*, dynlib.* python3.* into src/Callback/
Group src files doing similar things together
