+++ src/swaystatus.c
@@ -19,7 +19,7 @@
 #include "help.h"
 #include "utility.h"
 #include "printer.hpp"
-#include "python3.hpp"
+#include "Callback/python3.hpp"
 #include "process_configuration.h"
 #include "poller.h"
